+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Implement FindMergeBase RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/477
 - Update vendored gitlab_git to 359b65beac43e009b715c2db048e06b6f96b0ee8
   https://gitlab.com/gitlab-org/gitaly/merge_requests/481